+++ src/main/java/gila/notification/adapters/dto/request/CreateNotificationDTO.java
@@ -6,7 +6,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record CreateNotificationDTO (
-         @NotNull Long userId,
          @NotNull CategoryType category,
          @NotBlank String message
 ){}
