+++ src/main/java/gila/notification/application/gateways/NotificationGatewayImpl.java
@@ -1,6 +1,6 @@
 package gila.notification.application.gateways;
 
-import gila.notification.application.interfaces.gateways.NotificationGateway;
+import gila.notification.domain.interfaces.gateways.NotificationGateway;
 import gila.notification.domain.entities.Notification;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
