+++ src/main/java/gila/notification/application/usecases/CreateNotificationUseCaseImpl.java
@@ -1,6 +1,6 @@
 package gila.notification.application.usecases;
 
-import gila.notification.application.interfaces.usecases.CreateNotificationUseCase;
+import gila.notification.domain.interfaces.usecases.CreateNotificationUseCase;
 import gila.notification.domain.entities.Notification;
 
 public class CreateNotificationUseCaseImpl implements CreateNotificationUseCase {
