+++ src/main/java/gila/notification/infrastructure/beans/UseCaseBeanDeclaration.java
@@ -1,7 +1,7 @@
 package gila.notification.infrastructure.beans;
 
-import gila.notification.application.interfaces.usecases.CreateNotificationUseCase;
-import gila.notification.application.interfaces.usecases.GetAllNotificationsPagedUseCase;
+import gila.notification.domain.interfaces.usecases.CreateNotificationUseCase;
+import gila.notification.domain.interfaces.usecases.GetAllNotificationsPagedUseCase;
 import gila.notification.application.usecases.CreateNotificationUseCaseImpl;
 import gila.notification.application.usecases.GetAllNotificationsPagedUseCaseImpl;
 import org.springframework.context.annotation.Bean;
